feat(veil): introduce Pixelate effect and backward-compatible Blur
Introduces a new `Pixelate` effect (not stable yet) , which leverages `RuntimeShader` on API 33+ to apply a pixelation effect to the content. The size of the pixels is configurable.

This commit also refactors the `Blur` effect to be backward-compatible (may not work correctly) :
- On API 31+, it continues to use the efficient, built-in `Modifier.blur`.
- On older APIs (21-30), it falls back to a performant CPU-based `StackBlur` implementation. This involves taking a snapshot of the composable's window region, downscaling it, applying the blur, and then drawing it as an overlay.

The `Frosted` effect has been removed as it can be manually composed by combining `Blur` and `Scrim`. The sample app has been updated to reflect these changes.